rhi: Fix up the instancing step rate type
This is a UINT/uint32_t/GLuint/NSUInteger in all APIs (with Metal
being special due to being 64-bit in 64-bit apps whereas all others
are 32-bit always, at least on 64-bit Windows)

As the stride is already an uint32, follow suit for the step rate.
There was no reason to have this as int in the first place.

As an added bonus, some casts, that were previously needed when
mapping to the underlying API reality, can now be removed.

+++ src/gui/rhi/qrhid3d11.cpp
@@ -4425,7 +4425,7 @@ bool QD3D11GraphicsPipeline::create()
             const QRhiVertexInputBinding *inputBinding = m_vertexInputLayout.bindingAt(it->binding());
             if (inputBinding->classification() == QRhiVertexInputBinding::PerInstance) {
                 desc.InputSlotClass = D3D11_INPUT_PER_INSTANCE_DATA;
-                desc.InstanceDataStepRate = UINT(inputBinding->instanceStepRate());
+                desc.InstanceDataStepRate = inputBinding->instanceStepRate();
             } else {
                 desc.InputSlotClass = D3D11_INPUT_PER_VERTEX_DATA;
             }
